// amplify/types/riot/match-v5.ts
export interface MatchDto {
  /** Match metadata */
  metadata: MatchMetadataDto;
  /** Match info */
  info: MatchInfoDto;
}

/**
 * Match Metadata DTO
 */
export interface MatchMetadataDto {
  /** Match data version */
  dataVersion: string;
  /** Match ID */
  matchId: string;
  /** A list of participant PUUIDs */
  participants: string[];
}

/**
 * Match Info DTO
 */
export interface MatchInfoDto {
  /** Unix timestamp for when the game is created on the game server (i.e., the loading screen). */
  gameCreation: number;
  /** Prior to patch 11.20, this field returns the game length in milliseconds calculated from gameEndTimestamp - gameStartTimestamp.
   * Post patch 11.20, this field returns the max timePlayed of any participant in the game in seconds, which makes the behavior of this field consistent with that of match-v4.
   * The best way to get the exact game duration is to query the /timeline endpoint */
  gameDuration: number;
  /** Unix timestamp for when match ends on the game server. This timestamp can occasionally be significantly longer than when the match "ends".
   * The most reliable way of determining the timestamp for the end of the match would be to add the max time played of any participant to the gameStartTimestamp. */
  gameEndTimestamp: number;
  /** Game ID */
  gameId: number;
  /** Refer to the Game Constants documentation. */
  gameMode: string;
  /** Game name */
  gameName: string;
  /** Unix timestamp for when the game starts on the game server. */
  gameStartTimestamp: number;
  /** Game type */
  gameType: string;
  /** The first two parts can be used to determine the patch a game was played on. */
  gameVersion: string;
  /** Refer to the Game Constants documentation. */
  mapId: number;
  /** List of participants */
  participants: ParticipantDto[];
  /** Platform where the match was played. */
  platformId: string;
  /** Refer to the Game Constants documentation. */
  queueId: number;
  /** Teams information */
  teams: TeamDto[];
  /** Tournament code used to generate the match. This field is omitted if the match was not generated from a tournament code. */
  tournamentCode?: string;
}
// ...
export interface TeamDto {
  /** List of team bans */
  bans: BanDto[];
  /** Objectives information */
  objectives: ObjectivesDto;
  /** Team ID */
  teamId: number;
  /** Whether the team won */
  win: boolean;
}

/**
 * Ban DTO
 */
export interface BanDto {
  /** Champion ID */
  championId: number;
  /** Pick turn */
  pickTurn: number;
}

/**
 * Objectives DTO
 */
export interface ObjectivesDto {
  /** Baron objective information */
  baron: ObjectiveDto;
  /** Champion objective information */
  champion: ObjectiveDto;
  /** Dragon objective information */
  dragon: ObjectiveDto;
  /** Inhibitor objective information */
  inhibitor: ObjectiveDto;
  /** Rift Herald objective information */
  riftHerald: ObjectiveDto;
  /** Tower objective information */
  tower: ObjectiveDto;
}